requirements/base.txt
Django==2.1.5
djangorestframework==3.9.1
pytz
django-filter==2.1.0